tests/agent: add the ability to run queries during tests
And add a yelp test
