+++ languages/thingtalk/en/dlg/search-questions.genie
@@ -118,6 +118,7 @@ disjunctive_search_question = {
         return [pname];
     };
 
+    // FIXME we need to check that the base table matches the current domain
     ( ('are you looking for' | 'would you like' | 'do you need') 'a' table:base_table ('with' | 'that has') question:question_npp_filter_pair '?'
     | ('are you looking for' | 'would you like' | 'do you need') 'a' question:question_apv_filter_pair table:base_table '?'
     | ('are you looking for' | 'would you like' | 'do you need') 'a' table:base_table question:question_pvp_filter_pair '?'
@@ -142,7 +143,7 @@ simple_search_question = {
     one_param_direct_search_question;
     two_param_direct_search_question;
 
-    disjunctive_search_question;
+    !inference disjunctive_search_question;
 
     one_param_yesno_direct_search_question;
     two_param_yesno_direct_search_question;
