+++ lib/dialogue-agent/dialogue-loop.js
@@ -168,6 +168,29 @@ module.exports = class DialogueLoop {
         }
     }
 
+    async _handleUICommand(intent) {
+        switch (intent.type) {
+        case 'stop':
+            // stop means cancel, but without a failure message
+            throw new CancellationError();
+
+        case 'nevermind':
+            await this.reset();
+            throw new CancellationError();
+
+        case 'debug':
+            await this.reply("Current State:\n" + this._dialogueState.prettyprint());
+            break;
+
+        case 'wakeup':
+            // nothing to do
+            break;
+
+        default:
+            await this.fail();
+        }
+    }
+
     async _computePrediction(intent) {
         // handle all intents generated internally and by the UI:
         //
@@ -202,27 +225,8 @@ module.exports = class DialogueLoop {
         }
 
         if (intent.isUICommand) {
-            switch (intent.type) {
-            case 'stop':
-                // stop means cancel, but without a failure message
-                throw new CancellationError();
-
-            case 'nevermind':
-                await this.reset();
-                throw new CancellationError();
-
-            case 'debug':
-                await this.reply("Current State:\n" + this._dialogueState.prettyprint());
-                return null;
-
-            case 'wakeup':
-                // nothing to do
-                return null;
-
-            default:
-                await this.fail();
-                return null;
-            }
+            await this._handleUICommand(intent);
+            return null;
         }
 
         assert(intent.isDialogueState);
@@ -251,8 +255,8 @@ module.exports = class DialogueLoop {
             //console.log(this._dialogueState.prettyprint());
 
             [this._dialogueState, this._executorState] = await this._executor.execute(this._dialogueState, this._executorState);
-            //console.log(`Execution state:`);
-            //console.log(this._dialogueState.prettyprint());
+            this.debug(`Execution state:`);
+            this.debug(this._dialogueState.prettyprint());
 
             let policyPrediction, expect, utterance;
             if (this._useNeuralNLG()) {
@@ -448,7 +452,10 @@ module.exports = class DialogueLoop {
         await this.setExpected(expected);
         let intent = await this.nextIntent();
         while (!intent.isAnswer || intent.category !== expected) {
-            await this.fail();
+            if (intent.isUICommand)
+                await this._handleUICommand(intent);
+            else
+                await this.fail();
             intent = await this.nextIntent();
         }
         if (expected === ValueCategory.YesNo)
@@ -464,7 +471,10 @@ module.exports = class DialogueLoop {
             await this.replyChoice(i, choices[i]);
         let intent = await this.nextIntent();
         while (!intent.isMultipleChoiceAnswer) {
-            await this.fail();
+            if (intent.isUICommand)
+                await this._handleUICommand(intent);
+            else
+                await this.fail();
             intent = await this.nextIntent();
         }
         return intent.value;
