+++ lib/dialogue-agent/dialogue_policy.js
@@ -17,7 +17,7 @@ const I18n = require('../i18n');
 const SentenceGenerator  = require('../sentence-generator/generator');
 const TargetLanguages = require('../languages');
 
-const MAX_DEPTH = 9;
+const MAX_DEPTH = 7;
 const TARGET_PRUNING_SIZE = 100;
 
 module.exports = class DialoguePolicy {
@@ -46,7 +46,7 @@ module.exports = class DialoguePolicy {
             schemaRetriever: manager.schemas,
             maxDepth: MAX_DEPTH,
             targetPruningSize: TARGET_PRUNING_SIZE,
-            debug: true,
+            debug: 1,
             contextInitializer(state, functionTable) {
                 // ask the target language to extract the constants from the context
                 const constants = target.extractConstants(state);
