+++ lib/languages/thingtalk/constants.js
@@ -60,7 +60,7 @@ function extractConstants(ast) {
                 break;
 
             default:
-                addConstant('GENERIC_ENTITY_' + value.type, value.display, value);
+                addConstant('GENERIC_ENTITY_' + value.type, value.display || value.value, value);
                 break;
             }
         }
