+++ test/agent/tests.txt
@@ -1,4 +1,4 @@
-# basic
+# 1-basic
 # Let's start simple...
 
 U: \t $dialogue @org.thingpedia.dialogue.transaction.execute;
@@ -11,13 +11,53 @@ A: >> expecting = raw_string
 U: \t $dialogue @org.thingpedia.dialogue.transaction.execute;
 U: now => @com.facebook.post(status="hello world");
 
-A: Ok, do you want me to post on facebook saying hello world?
+A: Ok, you want to post hello world on facebook. Is that it?
 A: >> context = $dialogue @org.thingpedia.dialogue.transaction.sys_confirm_action ; now => @com.facebook.post param:status = QUOTED_STRING_0 ; // {"QUOTED_STRING_0":"hello world"}
 A: >> expecting = yesno
 
 U: \t $dialogue @org.thingpedia.dialogue.transaction.execute;
 U: now => @com.facebook.post(status="hello world") #[confirm=enum(confirmed)];
 
-A: I post hello world on facebook for you. Anything else for you today?
+A: I post on facebook saying hello world for you.
 A: >> context = $dialogue @org.thingpedia.dialogue.transaction.sys_action_success ; now => @com.facebook.post param:status = QUOTED_STRING_0 #[ results = [ ] ] ; // {"QUOTED_STRING_0":"hello world"}
+A: >> expecting = null
+
+====
+# 2-yelp
+# simple yelp search
+U: \t $dialogue @org.thingpedia.dialogue.transaction.execute;
+U: now => @com.yelp.restaurant() => notify;
+
+A: I see str:ENTITY_com.yelp:restaurant::0: and str:ENTITY_com.yelp:restaurant::1:. Both are 2 star restaurants.
+A: >> context = $dialogue @org.thingpedia.dialogue.transaction.sys_recommend_two ; now => @com.yelp.restaurant => notify #[ results = [ { param:id = GENERIC_ENTITY_com.yelp:restaurant_0 , param:image_url = PICTURE_0 , param:link = URL_0 , param:cuisines = [ GENERIC_ENTITY_com.yelp:restaurant_cuisine_0 ] , param:price = enum:moderate , param:rating = 2 , param:reviewCount = 2 , param:geo = LOCATION_0 , param:phone = PHONE_NUMBER_0 } ] ] #[ count = NUMBER_0 ] #[ more = true ] ; // {"GENERIC_ENTITY_com.yelp:restaurant_0":{"value":"str:ENTITY_com.yelp:restaurant::0:","display":null},"PICTURE_0":"str:ENTITY_tt:picture::15:","URL_0":"str:ENTITY_tt:url::14:","GENERIC_ENTITY_com.yelp:restaurant_cuisine_0":{"value":"str:ENTITY_com.yelp:restaurant_cuisine::37:","display":null},"LOCATION_0":{"latitude":16,"longitude":13,"display":null},"PHONE_NUMBER_0":"str:ENTITY_tt:phone_number::49:","NUMBER_0":50}
+A: >> expecting = null
+
+# U: What price range is str:ENTITY_com.yelp:restaurant::0: ?
+U: \t $dialogue @org.thingpedia.dialogue.transaction.execute;
+U: now => [price] of @com.yelp.restaurant(), id == "str:ENTITY_com.yelp:restaurant::0:"^^com.yelp:restaurant => notify;
+
+A: Str:ENTITY_com.yelp:restaurant::0: is a restaurant that is moderate.
+A: >> context = $dialogue @org.thingpedia.dialogue.transaction.sys_recommend_one ; now => [ param:price ] of ( ( @com.yelp.restaurant ) filter param:id == GENERIC_ENTITY_com.yelp:restaurant_0 ) => notify #[ results = [ { param:id = GENERIC_ENTITY_com.yelp:restaurant_0 , param:price = enum:moderate } ] ] ; // {"GENERIC_ENTITY_com.yelp:restaurant_0":{"value":"str:ENTITY_com.yelp:restaurant::0:","display":null}}
+A: >> expecting = null
+
+# U: No, I'm looking for something expensive.
+U: \t $dialogue @org.thingpedia.dialogue.transaction.execute;
+U: now => @com.yelp.restaurant(), price == enum(expensive) => notify;
+
+A: I have found str:ENTITY_com.yelp:restaurant::3:. It is a 9 star restaurant with 3 reviews. Would you like to hear more about it?
+A: >> context = $dialogue @org.thingpedia.dialogue.transaction.sys_recommend_one ; now => ( @com.yelp.restaurant ) filter param:price == enum:expensive => notify #[ results = [ { param:id = GENERIC_ENTITY_com.yelp:restaurant_0 , param:image_url = PICTURE_0 , param:link = URL_0 , param:cuisines = [ GENERIC_ENTITY_com.yelp:restaurant_cuisine_0 ] , param:price = enum:expensive , param:rating = 9 , param:reviewCount = 3 , param:geo = LOCATION_0 , param:phone = PHONE_NUMBER_0 } ] ] #[ count = NUMBER_0 ] ; // {"GENERIC_ENTITY_com.yelp:restaurant_0":{"value":"str:ENTITY_com.yelp:restaurant::3:","display":null},"PICTURE_0":"str:ENTITY_tt:picture::45:","URL_0":"str:ENTITY_tt:url::14:","GENERIC_ENTITY_com.yelp:restaurant_cuisine_0":{"value":"str:ENTITY_com.yelp:restaurant_cuisine::12:","display":null},"LOCATION_0":{"latitude":14,"longitude":168,"display":null},"PHONE_NUMBER_0":"str:ENTITY_tt:phone_number::44:","NUMBER_0":26}
+A: >> expecting = command
+
+# U: No thanks, this is sufficient.
+U: \t $dialogue @org.thingpedia.dialogue.transaction.cancel;
+
+A: Anything else for you today?
+A: >> context = $dialogue @org.thingpedia.dialogue.transaction.sys_anything_else ; now => ( @com.yelp.restaurant ) filter param:price == enum:expensive => notify #[ results = [ { param:id = GENERIC_ENTITY_com.yelp:restaurant_0 , param:image_url = PICTURE_0 , param:link = URL_0 , param:cuisines = [ GENERIC_ENTITY_com.yelp:restaurant_cuisine_0 ] , param:price = enum:expensive , param:rating = 9 , param:reviewCount = 3 , param:geo = LOCATION_0 , param:phone = PHONE_NUMBER_0 } ] ] #[ count = NUMBER_0 ] ; // {"GENERIC_ENTITY_com.yelp:restaurant_0":{"value":"str:ENTITY_com.yelp:restaurant::3:","display":null},"PICTURE_0":"str:ENTITY_tt:picture::45:","URL_0":"str:ENTITY_tt:url::14:","GENERIC_ENTITY_com.yelp:restaurant_cuisine_0":{"value":"str:ENTITY_com.yelp:restaurant_cuisine::12:","display":null},"LOCATION_0":{"latitude":14,"longitude":168,"display":null},"PHONE_NUMBER_0":"str:ENTITY_tt:phone_number::44:","NUMBER_0":26}
+A: >> expecting = command
+
+# U: No thanks, goodbye
+U: \t $dialogue @org.thingpedia.dialogue.transaction.end;
+
+A: Alright, good bye!
+A: >> context = $dialogue @org.thingpedia.dialogue.transaction.sys_end ; now => ( @com.yelp.restaurant ) filter param:price == enum:expensive => notify #[ results = [ { param:id = GENERIC_ENTITY_com.yelp:restaurant_0 , param:image_url = PICTURE_0 , param:link = URL_0 , param:cuisines = [ GENERIC_ENTITY_com.yelp:restaurant_cuisine_0 ] , param:price = enum:expensive , param:rating = 9 , param:reviewCount = 3 , param:geo = LOCATION_0 , param:phone = PHONE_NUMBER_0 } ] ] #[ count = NUMBER_0 ] ; // {"GENERIC_ENTITY_com.yelp:restaurant_0":{"value":"str:ENTITY_com.yelp:restaurant::3:","display":null},"PICTURE_0":"str:ENTITY_tt:picture::45:","URL_0":"str:ENTITY_tt:url::14:","GENERIC_ENTITY_com.yelp:restaurant_cuisine_0":{"value":"str:ENTITY_com.yelp:restaurant_cuisine::12:","display":null},"LOCATION_0":{"latitude":14,"longitude":168,"display":null},"PHONE_NUMBER_0":"str:ENTITY_tt:phone_number::44:","NUMBER_0":26}
 A: >> expecting = null
